pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "techpulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dagster>=1.12.10",
    "dagster-webserver>=1.12.10",
    "dbt-duckdb>=1.10.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.7.0",
    "structlog>=25.1.0",
    "tenacity>=9.0.0",
    "uvicorn>=0.40.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend/src/techpulse"]

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
    "respx>=0.22.0",
    "ruff>=0.14.11",
    "vcrpy>=7.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["dagster.*", "dbt.*", "duckdb.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["backend/tests", "tools/compliance/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
